DOUBLE

+++ kdtree.cpp
@@ -40,16 +40,16 @@ void KdTree::create(const vector<KdTreeNode> &pre, const std::vector<KdTreeTemp>
     }
 
 //    int suml=0, summ=0, sumr=pre.size();
-//    float mi = 0.5*pre.size()+pow(max(suml, sumr), 2.0/3.0);
+//    double mi = 0.5*pre.size()+pow(max(suml, sumr), 2.0/3.0);
 //    split = com[w][0].x-eps;
 //    for(size_t i=0;i<com[w].size();)
 //    {
-//        float c = com[w][i].x;
+//        double c = com[w][i].x;
 //        for(;i<com[w].size() && com[w][i].x<c+eps;i++)
 //        {
 //            if(com[w][i].in)summ++, sumr--;else suml++, summ--;
 //        }
-//        float t = 0.5*abs(suml-sumr)+pow(max(suml, sumr), 2.0/3.0)+summ;
+//        double t = 0.5*abs(suml-sumr)+pow(max(suml, sumr), 2.0/3.0)+summ;
 //        if(t < mi)mi = t, split = c;
 //    }
 
@@ -95,25 +95,25 @@ void KdTree::create(const vector<KdTreeNode> &pre, const std::vector<KdTreeTemp>
     if(!vr.empty())r = new KdTree(vr, com, (w+1)%3);
 }
 
-bool KdTree::check_node(const std::vector<Object> &objs, const Ray &ray, int &no, int &nv, float &dis)
+bool KdTree::check_node(const std::vector<Object> &objs, const Ray &ray, int &no, int &nv, double &dis)
 {
     dis=1e20;
     for(size_t l=0;l<node.size();l++)
     {
         int i=node[l].no, j=node[l].nv;
-        float t, u, v;
+        double t, u, v;
         if((t=objs[i].polys[j].intersect(ray, u, v))<dis && t > eps)
             dis=t, no=i, nv=j;
     }
     return dis < 1e10;
 }
 
-bool KdTree::check(const std::vector<Object> &objs, const Ray &ray, int &no, int &nv, float &dis)
+bool KdTree::check(const std::vector<Object> &objs, const Ray &ray, int &no, int &nv, double &dis)
 {
     dis = 1e20;
-    float dis1=1e20;
+    double dis1=1e20;
     if(!aabb.check_aabb(ray))return false;
-    float s[3]={ray.o.x, ray.o.y, ray.o.z}, v[3]={ray.d.x, ray.d.y, ray.d.z};
+    double s[3]={ray.o.x, ray.o.y, ray.o.z}, v[3]={ray.d.x, ray.d.y, ray.d.z};
     int no1, nv1;
     bool t = false;
     if(s[w] < split)
