+++ sphere.h
@@ -4,16 +4,16 @@
 
 struct Sphere
 {
-    float rad;
+    double rad;
     Point3D pos, lig, col;
     RType type;
-    __device__ __host__ Sphere(float rad_, Point3D pos_, Point3D lig_, Point3D col_, RType type_)
+    __device__ __host__ Sphere(double rad_, Point3D pos_, Point3D lig_, Point3D col_, RType type_)
         :rad(rad_), pos(pos_), lig(lig_), col(col_), type(type_) {
     }
-    __device__ __host__ float intersect(const Ray & r) const
+    __device__ __host__ double intersect(const Ray & r) const
     {
         Point3D op = pos - r.o;
-        float t, b = op * r.d, det = b * b - op * op + rad * rad;
+        double t, b = op * r.d, det = b * b - op * op + rad * rad;
         if (det < 0)return 1e20;
         else
             det = sqrt(det);
@@ -21,6 +21,6 @@ struct Sphere
     }
 };
 
-bool sphere_intersect(const Sphere spheres[], int n, const Ray & r, int &id, float &t);
+bool sphere_intersect(const Sphere spheres[], int n, const Ray & r, int &id, double &t);
 
 #endif
